Move jaq-core into jaq-std.
