--- jaq-core/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "jaq-core"
-version = "1.5.1"
-authors = ["Michael Färber <michael.faerber@gedenkt.at>"]
-edition = "2021"
-license = "MIT"
-readme = "../README.md"
-description = "Default native filters for jaq"
-repository = "https://github.com/01mf02/jaq"
-keywords = ["json", "query", "jq"]
-rust-version = "1.63"
-
-[features]
-default = ["std", "format", "log", "math", "parse_json", "regex", "time"]
-std = []
-format = ["aho-corasick", "base64", "urlencoding"]
-math = ["libm"]
-parse_json = ["hifijson"]
-time = ["chrono"]
-
-[dependencies]
-jaq-interpret = { version = "1.5.0", path = "../jaq-interpret" }
-hifijson = { version = "0.2.0", optional = true }
-chrono = { version = "0.4.38", default-features = false, features = ["alloc"], optional = true }
-regex = { version = "1.9", optional = true }
-log = { version = "0.4.17", optional = true }
-libm = { version = "0.2.7", optional = true }
-aho-corasick = { version = "1.0", optional = true }
-base64 = { version = "0.22", optional = true }
-urlencoding = { version = "2.1.3", optional = true }
-
-[dev-dependencies]
-serde_json = "1.0"
