+++ jaq-std/Cargo.toml
@@ -8,11 +8,26 @@ readme = "../README.md"
 description = "Standard library for jaq"
 repository = "https://github.com/01mf02/jaq"
 keywords = ["json", "query", "jq"]
-rust-version = "1.64"
+rust-version = "1.63"
+
+[features]
+default = ["std", "format", "log", "math", "parse_json", "regex", "time"]
+std = []
+format = ["aho-corasick", "base64", "urlencoding"]
+math = ["libm"]
+parse_json = ["hifijson"]
+time = ["chrono"]
 
 [dependencies]
-jaq-interpret   = { version = "1.5.0", path = "../jaq-interpret" }
+jaq-interpret = { version = "1.5.0", path = "../jaq-interpret" }
+hifijson = { version = "0.2.0", optional = true }
+chrono = { version = "0.4.38", default-features = false, features = ["alloc"], optional = true }
+regex = { version = "1.9", optional = true }
+log = { version = "0.4.17", optional = true }
+libm = { version = "0.2.7", optional = true }
+aho-corasick = { version = "1.0", optional = true }
+base64 = { version = "0.22", optional = true }
+urlencoding = { version = "2.1.3", optional = true }
 
 [dev-dependencies]
-jaq-core      = { version = "1.4.0", path = "../jaq-core" }
 serde_json = "1.0"
