+++ jaq-std/src/math.rs
