+++ jaq-std/tests/common/mod.rs
@@ -5,10 +5,10 @@ fn yields(x: Val, code: &str, ys: impl Iterator<Item = ValR>) {
     use jaq_interpret::load::{Arena, File, Loader};
 
     let arena = Arena::default();
-    let loader = Loader::new(jaq_std::std());
+    let loader = Loader::new(jaq_std::defs());
     let modules = loader.load(&arena, File { path: "", code }).unwrap();
     let filter = jaq_interpret::Compiler::default()
-        .with_funs(jaq_core::core())
+        .with_funs(jaq_std::funs())
         .compile(modules)
         .unwrap();
     filter.yields(x, ys)
