+++ jaq/src/main.rs
@@ -267,7 +267,7 @@ fn parse(path: &str, code: &str, vars: &[String]) -> Result<(Vec<Val>, Filter), 
 
     let vars: Vec<_> = vars.iter().map(|v| format!("${v}")).collect();
     let arena = Arena::default();
-    let loader = Loader::new(jaq_std::std()).with_std_read();
+    let loader = Loader::new(jaq_std::defs()).with_std_read();
     let modules = loader
         .load(&arena, File { path, code })
         .map_err(load_errors)?;
@@ -280,7 +280,7 @@ fn parse(path: &str, code: &str, vars: &[String]) -> Result<(Vec<Val>, Filter), 
     .map_err(load_errors)?;
 
     let compiler = Compiler::default()
-        .with_funs(jaq_core::core())
+        .with_funs(jaq_std::funs())
         .with_global_vars(vars.iter().map(|v| &**v));
     let filter = compiler.compile(modules).map_err(compile_errors)?;
     Ok((vals, filter))
